+++ packages/es-markdown-core/src/index.ts
@@ -2,6 +2,7 @@ export { Document } from "./document.js";
 export type { BlockQuote } from "./model/block-quote.js";
 export type { CodeBlock } from "./model/code-block.js";
 export type { Code } from "./model/code.js";
+export type { Comment } from "./model/comment.js";
 export type { DirectiveAttribute } from "./model/directive-attribute.js";
 export type { DirectiveBlock } from "./model/directive-block.js";
 export type { DirectiveLabel } from "./model/directive-label.js";