+++ packages/es-markdown-core/tests/visitor.mjs
@@ -23,6 +23,7 @@ export const kindVisitor = {
     visitLineBreak: (node) => node.kind,
     visitCode: (node) => node.kind,
     visitHTML: (node) => node.kind,
+    visitComment: (node) => node.kind,
     visitFormula: (node) => node.kind,
     visitEmphasis: (node) => node.kind,
     visitStrong: (node) => node.kind,