+++ packages/swift-markdown-core/Sources/MarkdownCore/Visitor/MarkupVisitor.swift
@@ -28,6 +28,7 @@ public protocol MarkupVisitor {
     mutating func visit(_ node: LineBreak) -> Result
     mutating func visit(_ node: Code) -> Result
     mutating func visit(_ node: HTML) -> Result
+    mutating func visit(_ node: Comment) -> Result
     mutating func visit(_ node: Formula) -> Result
     mutating func visit(_ node: Emphasis) -> Result
     mutating func visit(_ node: Strong) -> Result